requirements.txt
google-generativeai==0.3.0
pydantic==2.5.0
pypdf==4.0.1
flask==3.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
gunicorn==21.2.0
